Disable Semantic Commit prefixes for Renovate
This was done because, by default, Renovate will try to autodetect if a
repository already uses Semantic Commits. Which, in turn, affects the
commit messages and PR titles that Renovate writes. Despite the
vscode-sync-settings extension writing commit messages in a Semantic
Commit-like form, it is not the style I currently write commit messages
in.